src/sage/libs/giac/giac.pyx: fix signed infinity handling
Before this commit, signed Sage infinities would be converted to
unsigned Giac infinities:

    sage: libgiac(+Infinity)
    Infinity

This can mess up your integrals, if nothing else. To fix it, we add an
additional case to the Pygen constructor. Sage's AnInfinity class
already knows how to convert itself to a Giac-friendly string, and
Pygen already knows what to do with those strings -- we just need
to put the pieces together. A new test case is added as well.